feat : nickname maker & S3 base
