+++ src/main/java/clone/gozik/GozikApplication.java
@@ -1,0 +1,13 @@
+package clone.gozik;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GozikApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GozikApplication.class, args);
+    }
+
+}
