+++ src/main/java/clone/gozik/entity/EmployeeType.java
@@ -1,0 +1,8 @@
+package clone.gozik.entity;
+
+public enum EmployeeType {
+    NEW,//신입
+    INTERN,//인턴
+    EXPERIENCED,//경력직
+    CONTRACT//계약직
+}
