+++ src/main/java/clone/gozik/entity/RecruitTypeEnum.java
@@ -1,0 +1,7 @@
+package clone.gozik.entity;
+
+public enum RecruitTypeEnum {
+    ONGOING,//상시
+    CLOSED,//마감
+    OPEN//채용중
+}
