+++ src/main/java/clone/gozik/exception/GGANTTONG11.java
@@ -1,0 +1,4 @@
+package clone.gozik.exception;
+
+public class GGANTTONG11 {
+}
