+++ src/main/java/clone/gozik/repository/BoardRepository.java
@@ -1,0 +1,8 @@
+package clone.gozik.repository;
+
+import clone.gozik.entity.Board;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BoardRepository extends JpaRepository<Board,Long> {
+
+}
