+++ src/main/java/clone/gozik/repository/JobRepository.java
@@ -1,0 +1,11 @@
+package clone.gozik.repository;
+
+import clone.gozik.entity.Board;
+import clone.gozik.entity.Job;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface JobRepository extends JpaRepository<Job,Long> {
+    List<Job> findByBoard(Board board);
+}
