+++ src/main/java/clone/gozik/security/GGAngtong47.java
@@ -1,0 +1,4 @@
+package clone.gozik.security;
+
+public class GGAngtong47 {
+}
