+++ src/main/java/clone/gozik/service/RandomNicknameBox.java
@@ -1,0 +1,62 @@
+package clone.gozik.service;
+
+
+public class RandomNicknameBox {
+
+
+
+    private String[] prefix = {
+            "불가능한","어질어질한","앙탈부리는","금쪽같은","아름다운",
+            "코드짜는","심심한","시시한","탐나는","질투심많은",
+            "리더십강한","손수만든","배고픈","배부른","혼자있고싶은",
+            "강력한","아직어린","냄새나는","향기로운","비싸보이는",
+            "바보같은","멍청한","양심없는","부유한","귀티나는",
+            "삐짝바른","뚱뚱한","무거운","가벼운","병약한",
+            "성스러운","밥먹으러온","약탈하러온","땡깡부리는","노래하는",
+            "당황하는","방황하는","짐짝같은","쓸모없는","벌레같은",
+            "디자인잘하는","개발잘하는","반짝이는","어두컴컴한","훌륭한",
+            "운수좋은","운빨개쩌는","보좌하는","싱싱한","낡아빠진",
+            "오래된","로또당첨안되는","통통튀는","엎어진","쓰러진",
+            "고통스러운","밥짓는","땅파는","삽질하는","오류덩어리인",
+            "우락부락한","건강한","KPOP듣는","애국심넘치는","행복한",
+            "불행한","우중충한","태풍같은","태양같은","자유로운",
+            "억압당한","성장하는","잔잔한","폭풍같은","파도같은",
+            "놀러나온","얄미운","귀여운","이쁜","돌머리인",
+            "정치질잘하는","탑베인하는","미드마이하는","떙깡부리는","골렘같은",
+            "탐관오리인","말귀못알아먹는","똑똑한","똑부러지는","비타민같은",
+            "쓴맛나는","차가운","따듯한","미적찌근한","도움안되는",
+            "갈무리하는","가만히서있는","앉은뱅이","말많은","청청한"
+    };
+
+    private String[] suffix = {
+            "원숭이","뱀","고양이","팬더","늑대",
+            "갈매기","독수리","강아지","개",
+            "곰","돼지","닭","고래","오리",
+            "도마뱀","개구리","사람","도룡뇽","병아리",
+            "프로그래머","개발자","왕","왕비","신하",
+            "마피아","성직자","마약상","시민","시장",
+            "프로도","제이지","춘식이","신","장군",
+            "페페","스파이더맨","펭귄","배트맨","슈퍼맨",
+            "대머리","아이언맨","앤트맨","늑대","오이",
+            "호박","오징어","샐러리","소","용사",
+            "집사","암살자","해병","아쎄이","외발이",
+            "잭스","지렁이","개미","콜라병","종이컵",
+            "물로켓","로켓","개미핥기","다람쥐","쥐",
+            "햄스터","생쥐","사자","보물상자","도둑",
+            "경찰","외판원","톨스토이","파우스트","정치인",
+            "국회의원","대통령","대총통","커플","솔로",
+            "모태솔로","선생님","용","커피","껌",
+            "초콜랫","비타민","얼음","불","말",
+            "햄버거","샌드위치","돌","태양","김밥",
+            "라면","짜장면","짬뽕","탕수육","염소"
+    };
+
+
+    public String GetPrefix(){
+        return prefix[(int)(Math.random()*prefix.length)];
+    }
+
+    public String GetSuffix(){
+        return suffix[(int)(Math.random()*prefix.length)];
+    }
+}
