--- src/main/java/com/example/gozik/GozikApplication.java
@@ -1,13 +1,0 @@
-package com.example.gozik;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GozikApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(GozikApplication.class, args);
-    }
-
-}
