--- src/main/java/com/example/gozik/dto/RequestBoardDto.java
@@ -1,17 +1,0 @@
-package clone.gozic.dto;
-
-import clone.gozic.entity.CompanyTypeEnum;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-public class RequestBoardDto {
-    private String title;
-    private String description;
-    private CompanyTypeEnum companytype;
-    private String startDate;
-    private String endDate;
-    private boolean recruitmentPeriod;
-    private List job;
-}
