--- src/main/java/com/example/gozik/entity/EmployeeType.java
@@ -1,8 +1,0 @@
-package clone.gozic.entity;
-
-public enum EmployeeType {
-    NEW,//신입
-    INTERN,//인턴
-    EXPERIENCED,//경력직
-    CONTRACT//계약직
-}
