--- src/main/java/com/example/gozik/entity/Favorites.java
@@ -1,30 +1,0 @@
-package clone.gozic.entity;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor
-@Getter
-public class Favorites {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column
-    private int like;
-
-    @ManyToOne
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne
-    @JoinColumn(name = "board_id", nullable = false)
-    private Board board;
-
-//    public Favorites(responseDto){
-//
-//    }
-}
