--- src/main/java/com/example/gozik/entity/MemberRoleEnum.java
@@ -1,22 +1,0 @@
-package clone.gozic.entity;
-
-public enum MemberRoleEnum {
-
-    MEMBER(Authority.MEMBER),
-    COMPANY(Authority.COMPANY);
-
-    private final String authority;
-
-    MemberRoleEnum(String authority) {
-        this.authority = authority;
-    }
-
-    public String getAuthority(){
-        return authority;
-    }
-
-    public static class Authority{
-        public static final String MEMBER = "ROLE_MEMBER";
-        public static final String COMPANY = "ROLE_COMPANY";
-    }
-}
