--- src/main/java/com/example/gozik/entity/RecruitTypeEnum.java
@@ -1,7 +1,0 @@
-package clone.gozic.entity;
-
-public enum RecruitTypeEnum {
-    ONGOING,//상시
-    CLOSED,//마감
-    OPEN//채용중
-}
