--- src/main/java/com/example/gozik/exception/GGANTTONG11.java
@@ -1,4 +1,0 @@
-package clone.gozic.exception;
-
-public class GGANTTONG11 {
-}
