--- src/main/java/com/example/gozik/repository/BoardRepository.java
@@ -1,10 +1,0 @@
-package clone.gozic.repository;
-
-import clone.gozic.entity.Board;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface BoardRepository extends JpaRepository<Board,Long> {
-
-}
