--- src/main/java/com/example/gozik/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package clone.gozic.repository;
-
-import clone.gozic.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
-    Optional<Member> findByKakaoId(Long id);
-
-}
