--- src/main/java/com/example/gozik/security/GGAngtong47.java
@@ -1,4 +1,0 @@
-package clone.gozic.security;
-
-public class GGAngtong47 {
-}
