+++ src/test/java/clone/example/gozik/GozikApplicationTests.java
@@ -1,0 +1,13 @@
+package clone.example.gozik;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GozikApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
