--- src/test/java/com/example/gozik/GozikApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.gozik;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GozikApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
